Reworked way that trials/potentials are specified
Now all trails and potentials of a particular
kind export a uniform interface and all have the
same module name.  The choice of the trial and
potential is done at link time, where the object
files for the desired trials and potentials are
specified and linked to the rest of the program.