Moved potential and probability computation into separate func
Now two calls are made to this function, rather than copying
and pasting the code.

Also changed "acc_flag" to "reject_flag", since this more
clearly conveys how it is used (i.e., since one would expect
the normal case to be for a flag not to be flown, but "acc_flag"
was flown by default rather than by exception).