Validated that ang momentum cancels ang velocity!!
That is, I wrote a test that demonstrated that the
effective potential contribution from angular
momentum terms is lowest when the angular momentum
of the system is as close as possible to the
frame angular velocity.

It turned out that there were two keys to making
this work:  recognizing that my name for the
"angular momentum" variable was misleading, as
the true angular momentum is number of excited
particles divided by the total number of
particles, and including an extra factor of 1/2
on the kinetic energy term.