fixed_angular_momentum => N_rotating_particles
I renamed this variable in order to improve
clarity.  The true angular momentum is actually
not equal to the number of rotating particles, but
is actually equal to the number of rotating
particles divided by the total number of
particles.  This distinction is incredibly
important to make because getting it wrong means
that the effects of a rotating frame and rotating
particles within the frame don't cancel each
other out as one would physically expect.